refactoring ExecutionScore to be a type alias
