+++ src/main/scala/org/allenai/pnp/Pnp.scala
@@ -1,14 +1,14 @@
 package org.allenai.pnp
 
-import scala.collection.JavaConverters._
 import com.google.common.base.Preconditions
 import com.jayantkrish.jklol.training.LogFunction
 import com.jayantkrish.jklol.training.NullLogFunction
 import com.jayantkrish.jklol.util.CountAccumulator
+
+import ExecutionScore.ExecutionScore
 import edu.cmu.dynet._
+import edu.cmu.dynet.DyNetScalaHelpers.floatVectorToSeq
 import edu.cmu.dynet.dynet_swig._
-import edu.cmu.dynet.DyNetScalaHelpers._
-import scala.collection.mutable.MapBuilder
 
 /** Probabilistic neural program monad. Pnp[X] represents a
   * function from neural network parameters to a probabilistic
