+++ src/main/scala/org/allenai/pnp/semparse/SemanticParser.scala
@@ -10,7 +10,7 @@ import scala.collection.mutable.SetBuilder
 
 import org.allenai.pnp.CompGraph
 import org.allenai.pnp.Env
-import org.allenai.pnp.ExecutionScore
+import org.allenai.pnp.ExecutionScore.ExecutionScore
 import org.allenai.pnp.Pnp
 import org.allenai.pnp.Pnp._
 import org.allenai.pnp.PnpModel
