+++ src/test/scala/org/allenai/pnp/GlobalLoglikelihoodTrainerSpec.scala
@@ -1,11 +1,13 @@
 package org.allenai.pnp
 
-import scala.collection.JavaConverters._
-import org.scalatest._
+import org.allenai.pnp.ExecutionScore.ExecutionScore
+import org.scalatest.FlatSpec
+import org.scalatest.Matchers
+
+import com.jayantkrish.jklol.training.NullLogFunction
+
 import edu.cmu.dynet._
 import edu.cmu.dynet.dynet_swig._
-import com.jayantkrish.jklol.util.IndexedList
-import com.jayantkrish.jklol.training.NullLogFunction
 
 class GlobalLoglikelihoodTrainerSpec extends FlatSpec with Matchers {
